Update GameObject.h
CollisionInfo changed to return float type

+++ CSC8503CoreClasses/GameObject.h
@@ -84,7 +84,7 @@ namespace NCL::CSC8503 {
 		//unsigned int texture;
 		bool isPaintable;
 		
-		virtual int collisionInfo() {
+		virtual float collisionInfo() {
 			return 0;
 		}
